Cleaned up autoconf, etc., so it compiles. 	modified:   libintense-io/Makefile.am 	modified:   libintense-io/Makefile.in 	modified:   libintense-io/aclocal.m4 	modified:   libintense-io/autom4te.cache/output.0 	modified:   libintense-io/autom4te.cache/output.1 	new file:   libintense-io/autom4te.cache/output.2 	modified:   libintense-io/autom4te.cache/requests 	modified:   libintense-io/autom4te.cache/traces.0 	modified:   libintense-io/autom4te.cache/traces.1 	new file:   libintense-io/autom4te.cache/traces.2 	modified:   libintense-io/config/ltmain.sh 	new file:   libintense-io/config/test-driver 	modified:   libintense-io/configure 	modified:   libintense-io/configure.ac 	modified:   libintense-io/include/Makefile.in 	modified:   libintense-io/include/intense-io/XDRBaseSerialiser.hpp 	new file:   libintense-io/m4/libtool.m4 	new file:   libintense-io/m4/ltoptions.m4 	new file:   libintense-io/m4/ltsugar.m4 	new file:   libintense-io/m4/ltversion.m4 	new file:   libintense-io/m4/lt~obsolete.m4 	modified:   libintense-io/src/BinaryBaseSerialiser.cpp 	modified:   libintense-io/src/LogStream.cpp 	modified:   libintense-io/src/Makefile.am 	modified:   libintense-io/src/Makefile.in 	modified:   libintense-io/src/XDRBaseSerialiser.cpp 	modified:   libintense-io/tests/BinarySerialiserTestSource.cpp 	modified:   libintense-io/tests/Makefile.am 	modified:   libintense-io/tests/Makefile.in 	modified:   libintense-io/tests/find_compatible_arch.sh

+++ libintense-io/include/intense-io/XDRBaseSerialiser.hpp
@@ -85,16 +85,16 @@ namespace intense {
       (short unsigned int& i, std::istream& is);
 
       virtual void out
-      (long long int i, std::ostream& os);
+      (long int i, std::ostream& os);
 
       virtual void in
-      (long long int& i, std::istream& is);
+      (long int& i, std::istream& is);
 
       virtual void out
-      (unsigned long long int i, std::ostream& os);
+      (unsigned long int i, std::ostream& os);
 
       virtual void in
-      (unsigned long long int& i, std::istream& is);
+      (unsigned long int& i, std::istream& is);
 
       void out
       (char c, std::ostream& os);
